Task fixed correctly
