+++ src/main/java/acme/entities/sessions/Tutorial.java
@@ -1,31 +1,36 @@
 
-package acme.datatypes.sessions;
+package acme.entities.sessions;
 
-import java.util.Date;
-
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
-import org.hibernate.validator.constraints.URL;
 
 import acme.framework.data.AbstractEntity;
+import acme.roles.Assistant;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class sessions extends AbstractEntity {
+public class Tutorial extends AbstractEntity {
 
 	// Serialisation identifier -----------------------------------------------
 
 	protected static final long	serialVersionUID	= 1L;
 
 	// Attributes -------------------------------------------------------------
+
+	@Column(unique = true)
+	@NotBlank
+	@Pattern(regexp = "[A-Z]{1,3}[0-9][0-9]{3}", message = "{validation.tutorial.code}")
+	protected String			code;
 
 	@NotBlank
 	@Length(min = 0, max = 75)
@@ -35,20 +40,22 @@ public class sessions extends AbstractEntity {
 	@Length(min = 0, max = 100)
 	protected String			resume;
 
-	protected SessionType		sessionType;
+	@NotBlank
+	@Length(min = 0, max = 100)
+	protected String			goals;
 
-	protected Date				periodStart;
-
-	protected Date				periodEnd;
-
-	@URL
-	protected String			information;
+	protected Double			estimatedTotalTime;
 
 	// Relationships ----------------------------------------------------------
 
 	@NotNull
 	@Valid
 	@ManyToOne(optional = false)
-	protected Tutorial			tutorial;
+	protected Assistant			assistant;
+
+	//@NotNull
+	//@Valid
+	//@ManyToOne(optional = false)
+	//protected course			course;
 
 }
