+++ src/main/java/acme/entities/sessions/TutorialSession.java
@@ -1,36 +1,31 @@
 
-package acme.datatypes.sessions;
+package acme.entities.sessions;
 
-import javax.persistence.Column;
+import java.util.Date;
+
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
+import org.hibernate.validator.constraints.URL;
 
 import acme.framework.data.AbstractEntity;
-import acme.roles.Assistant;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class Tutorial extends AbstractEntity {
+public class TutorialSession extends AbstractEntity {
 
 	// Serialisation identifier -----------------------------------------------
 
 	protected static final long	serialVersionUID	= 1L;
 
 	// Attributes -------------------------------------------------------------
-
-	@Column(unique = true)
-	@NotBlank
-	@Pattern(regexp = "[A-Z]{1,3}[0-9][0-9]{3}", message = "{validation.tutorial.code}")
-	protected String			code;
 
 	@NotBlank
 	@Length(min = 0, max = 75)
@@ -40,22 +35,20 @@ public class Tutorial extends AbstractEntity {
 	@Length(min = 0, max = 100)
 	protected String			resume;
 
-	@NotBlank
-	@Length(min = 0, max = 100)
-	protected String			goals;
+	protected SessionType		sessionType;
 
-	protected Double			estimatedTotalTime;
+	protected Date				periodStart;
+
+	protected Date				periodEnd;
+
+	@URL
+	protected String			information;
 
 	// Relationships ----------------------------------------------------------
 
 	@NotNull
 	@Valid
 	@ManyToOne(optional = false)
-	protected Assistant			assistant;
-
-	//@NotNull
-	//@Valid
-	//@ManyToOne(optional = false)
-	//protected course			course;
+	protected Tutorial			tutorial;
 
 }
