+++ src/main/webapp/WEB-INF/resources/sample-data/assistant.csv
@@ -1,0 +1,3 @@
+key,supervisor,expertiseFields,resume,furtherInformation,key:userAccount
+assistant-01,"Pedro Jose Giraldo","Biology,physics...","Example of college degree","http://www.example.com/",user-account-assistant-01
+assistant-02,"Marcelino Olivera","Example 1,Example 2","Example of college degree","http://www.example.com/",user-account-assistant-02
