+++ src/main/webapp/WEB-INF/resources/sample-data/banner.csv
@@ -1,0 +1,3 @@
+key,moment,displayPeriodStart,displayPeriodEnd,linkPicture,slogan,linkTarget
+banner-01,2021/02/13 19:39,2021/02/14 19:39,2021/02/14 20:39,"http://www.example.com/","Example","http://www.example.com/"
+banner-02,2022/02/13 19:39,2022/02/14 19:39,2022/02/14 20:39,"http://www.example.com/","Example","http://www.example.com/"
