+++ src/main/webapp/WEB-INF/resources/sample-data/tutorial.csv
@@ -1,0 +1,3 @@
+key,code,title,resume,goals,estimatedTotalTime,key:assistant
+tutorial-01,BBB0000,"Title one","Summary 1","Some goals",1.0,assistant-01
+tutorial-02,AAA0000,"Title two","Summary 1","Some goals",1.5,assistant-01
