+++ src/main/webapp/WEB-INF/resources/sample-data/user-account.csv
@@ -1,0 +1,3 @@
+key,enabled,identity,password,username
+user-account-assistant-01,true,"UserIdentity{name: 'Assistant1', surname: 'Acme', email: 'assistant1@acme.com'}",assistant1,assistant1
+user-account-assistant-02,true,"UserIdentity{name: 'Assistant2', surname: 'Acme', email: 'assistant2@acme.com'}",assistant2,assistant2
